+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Basé sur [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - API MQ Delibs: les annexes renvoient leurs content-type
+- S'assure que le champ `objet` des metadata pastell est bien encodé en `latin-1`
 
 ## [2.0.11] - 2022-12-07